Refactor F/A-18 PrePlannedBuilder class
The stations are grouped by payload type by GroupStationsByPayloadType
into a List<List<PrePlannedStation>>.

The outer List contains the groups, which are in the order by which the
payload type was first encountered while iterating through stations in
order 2,3,7,8 (This matches the order as payload types are listed on the
MFD).

The inner List contains the individual stations of that group, which are
ordered as 8,2,7,3 (stepping order of stations with the same payload).

Other changes:
 - Changed InputStation to only press PP1 once at the end to return to
   it if needed.
 - Renamed isJdam to isJdamOrJsow
 - Simplified switch-case structure in GetDsplyCommand and isJdamOrJsow
   for readability.